pyproject.toml
[tool.poetry]
name = "csu-class-bot"
version = "0.1.0.alpha"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0-alpha.16"
nonebot-adapter-telegram = "^0.1.0-alpha.0"
beautifulsoup4 = "^4.10.0"
prettytable = "^2.4.0"
pillow = "^8.4.0"

[tool.poetry.dev-dependencies]
nb-cli = "^0.5.1"
black = { version = "^21.10b0", allow-prereleases = true }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
